backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
plotly
phidata
duckduckgo-search
yfinance
groq